Configuration file

+++ update.php
@@ -21,54 +21,40 @@
 		<p>
 <?php
 
-if (!empty($_POST["url"])) {
-	$table = $_POST["team"];
-	$url = "https://wt-90a32ef8ce619cb1158f5b6850fb1f12-0.sandbox.auth0-extend.com/maxprepsTeamScraper?url=" . $_POST["url"]; // path to the JSON file
-	$data = file_get_contents($url);  // load contents of the page 
-	$roster = json_decode($data, true); // parse the JSON into an object
-	
-    $output = "INSERT INTO $table (Number, Name, Position) VALUES ";
-	
-	foreach ($roster as $player) {
-        if ($player[jersey]){
-            $output = $output . "('" . $player[jersey] . "','" . $player[name] . "','" . $player[position] . "'), ";
-        }
-	   
+	if (!empty($_POST["url"])) {
+		$table = $_POST["team"];
+		$url = "https://wt-90a32ef8ce619cb1158f5b6850fb1f12-0.sandbox.auth0-extend.com/maxprepsTeamScraper?url=" . $_POST["url"]; // path to the JSON file
+		$data = file_get_contents($url);  // load contents of the page 
+		$roster = json_decode($data, true); // parse the JSON into an object
+		
+		// Build insert statement
+		$output = "INSERT INTO $table (Number, Name, Position) VALUES ";
+		foreach ($roster as $player) {
+			if ($player[jersey]){
+				$output = $output . "('" . $player[jersey] . "','" . $player[name] . "','" . $player[position] . "'), ";
+			}
+
+		}
+		$output = $output . "(null,null,null);";
+
+		// Load config values
+		include 'config.php';
+		if ($host_name == "") {
+			die("<h3>You need to make a copy of the 'config.example.php' and name it 'config.php', then edit that file with the appropriate database access parameters.</h3>");
+		}
+		
+		$conn = mysqli_connect($host_name, $user_name, $password, $database);
+		if (mysqli_connect_errno()) {
+		  die('<p>Failed to connect to MySQL: '.mysqli_connect_error().'</p>');
+		} 
+		else
+
+		{mysqli_query($conn, "DELETE FROM $table;");
+		if (mysqli_query($conn, $output) == false) {echo "error" . (mysqli_error($conn));}
+		 // Execute the insert statement
+		mysqli_close($conn);
+		}
 	}
-	$output = $output . "(null,null,null);";
-//	echo "<pre>" . $output . "</pre>";
-// Open DB
-    
-    function readCSV($csvFile){
-        $file_handle = fopen($csvFile, 'r');
-        while (!feof($file_handle) ) {
-            $line_of_text[] = fgetcsv($file_handle, 1024);
-        }
-        fclose($file_handle);
-        return $line_of_text;
-    }
-    $csvFile = '../.sec/hold';
-
-    $csv = readCSV($csvFile);
-    $host_name = $csv[1][0];
-    $database = $csv[1][1];
-    $user_name = $csv[1][2];
-    $password = $csv[1][3];
-
-    fclose($handle);
-
-	$conn = mysqli_connect($host_name, $user_name, $password, $database);
-    if (mysqli_connect_errno()) {
-      die('<p>Failed to connect to MySQL: '.mysqli_connect_error().'</p>');
-    } 
-    else
-        
-	{mysqli_query($conn, "DELETE FROM $table;");
-    if (mysqli_query($conn, $output) == false) {echo "error" . (mysqli_error($conn));}
-	 // Execute the insert statement
-	mysqli_close($conn);
-    }
-}
 
 ?></p>
 	</body>
